🚸 tqec is not the main CLI and add check_dae subcommand
