+++ src/tqec/_cli/subcommands/base.py
@@ -1,0 +1,16 @@
+import argparse
+from abc import ABC, abstractmethod
+
+
+class TQECSubCommand(ABC):
+    """Interface that should be implemented by subcommands of the tqec CLI."""
+
+    @staticmethod
+    @abstractmethod
+    def add_subcommand(main_parser: argparse.ArgumentParser) -> None:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def execute(args: argparse.Namespace) -> None:
+        pass
